micro-api-countries/src/main/java/com/afym/model/Country.java
package com.afym.model;

import java.util.List;

public class Country {
    private String name;
    private String flag;
    private List<City> cities;
    private List<Population> populations;

    public Country(String name){
        this.name = name;
        this.flag = "";
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getFlag() {
        return flag;
    }

    public void setFlag(String flag) {
        this.flag = flag;
    }

    public List<City> getCities() {
        return cities;
    }

    public void setCities(List<City> cities) {
        this.cities = cities;
    }

    public List<Population> getPopulations() {
        return populations;
    }

    public void setPopulations(List<Population> populations) {
        this.populations = populations;
    }
}
